CA-116058: Improve exception handling when resource is inaccessible
Raise a RuntimeError when a resource is not accessible so that xenopsd
handles the error gracefully and shows a message such as:

The bootloader returned an error
vm: 05ea81ab-b7db-dbca-c639-9ed35ed91ff1 (debhttp3)
msg: Could not access http://...

instead of showing an internal error.

Additionally, log the original exception for more useful debugging.