CA-220466: eliloader: use simple-format, not sxp
Emit simple-format, tell pygrub to emit simple-format, and
parse simple-format from pygrub.

This is to defend against an attack that is possible from the guest
when the s-expression output format is used: a guest can modify its
grub config kernel-line in such a way that when it boots, a dom0
file of the guest's choosing will be made available to the guest as
a ramdisk, and then deleted.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>